Melee Weapon Item Created

+++ Tundra3/Frostbyte/scripts/ItemSystem/ItemSystem.gml
@@ -87,38 +87,69 @@ function ConsumableItem(_name, _durability, _staminaCost, _weight, _consumeEffec
     }
 }
 
-function WeaponItem(_name, _weaponType, _durability, _staminaCost, _weight, _attackEffects) : Item(_name, _durability, _staminaCost, _weight, ItemType.Weapon, _attackEffects, true) constructor {
+function WeaponItem(_name, _weaponType, _durability, _staminaCost, _weight, _equipEffects, _attackEffects) : Item(_name, _durability, _staminaCost, _weight, ItemType.Weapon, _equipEffects, true) constructor {
     // get ref to owner
     // get ref to stat system
     // apply effects to stat system on equip
     weaponType = _weaponType;
+    stats = owner.stats;
 
     function GetWeaponType()
     {
         return weaponType;
     }
+
+    function Equip()
+    {
+        for(var i = 0; i < array_length(effects); i++) {
+            stats.AddStat(effects[i]);
+        }
+    }
+
+    function Unequip()
+    {
+        for(var i = 0; i < array_length(effects); i++) {
+            stats.RemoveStat(effects[i]);
+        }
+    }
 }
 
-function MeleeWeaponItem(_name, _durability, _staminaCost, _weight, _attackEffects) : WeaponItem(_name, WeaponType.Melee, _durability, _staminaCost, _weight, _attackEffects) constructor {
+function MeleeWeaponItem(_name, _color, _durability, _damage, _damageType, _staminaCost, _weight,  _equipEffects, _attackEffects) : WeaponItem(_name, WeaponType.Melee, _durability, _staminaCost, _weight, _equipEffects) constructor {
     // get ref to owner
     // get ref to stat system
     // apply effects to stat system on equip
+    damage = _damage;
+    damageType = _damageType;
+    color = _color;
+
+    function Use()
+    {
+        var meleeSwing = new BulletItem($"{name}'s Swing", true, 1, damage, damageType, 0, []);
+        meleeSwing.owner = self.owner;
+        
+        var meleeSwingObject = instance_create_layer(owner.x, owner.y, "MeleeSwing", obj_Bullet);
+        meleeSwingObject.SetBulletData(meleeSwing, 0, owner.targets, owner.image_angle);
+        meleeSwingObject.sprite_index = spr_meleeSlash;
+        meleeSwingObject.image_blend = color;
+    }
 }
 
-function RangedWeaponItem(_name, _durability, _staminaCost, _weight, _attackEffects) : WeaponItem(_name,, WeaponType.Ranged, _durability, _staminaCost, _weight, _attackEffects) constructor {
+function RangedWeaponItem(_name, _durability, _speed, _staminaCost, _weight,  _equipEffects, _attackEffects) : WeaponItem(_name,, WeaponType.Ranged, _durability, _staminaCost, _weight, _equipEffects) constructor {
     // get ref to owner
     // get ref to stat system
     // apply effects to stat system on equip
+    speed = _speed;
 }
 
-function BulletItem(_name, _speed, _damage, _damageType, _weight, _hitFunctions) : Item(_name, -1, 0, _weight, ItemType.Bullet, [], true) constructor {
+function BulletItem(_name, _hitMultiple, _lifeTime, _damage, _damageType, _weight, _hitFunctions) : Item(_name, -1, 0, _weight, ItemType.Bullet, [], true) constructor {
     // get ref to owner
     // get ref to stat system
     // apply effects to stat system
-    speed = _speed;
     damage = _damage;
     damageType = _damageType;
     hitFunctions = _hitFunctions;
+    lifeTime = _lifeTime * game_get_speed(gamespeed_fps);
+    hitMultiple = _hitMultiple;
 
     function Use(_position)
     {
@@ -155,6 +186,9 @@ function ArmorItem(_name, _armorValue, _durability, _staminaCost, _weight, _defe
     {
         durability -= 1;
         stamina.UseStamina(staminaCost);
+        if (stamina.GetCurrentStamina() <= 0) {
+            return 0;
+        }
 
         if (durability <= 0) {
             owner.inventory.DeleteItem(self, 1);
